requirements.txt
elasticsearch==8.15.1
pdfplumber==0.11.4
python-dotenv==1.0.1
pandas==2.2.3
openai==1.53.0
